Migrate fjs/sul/level/hash to .f.mjs; rename byte_set's Byte to _Byte
Splits the Add/EncodeState types into a sibling types.ts, and converts
the runtime implementation to JSDoc-typed module.f.mjs/proof.f.mjs
under the stage-1 TypeScript-to-mjs migration. Updates the two
dependents (fjs/sul/module.f.ts, fjs/sul/proof.f.ts) to import runtime
values from module.f.mjs and the Add type from types.ts.

Also renames fjs/types/byte_set/module.f.mjs's private Byte typedef to
_Byte: JSDoc typedefs are always exported, so the unprefixed name
leaked into the emitted declaration where main kept it private,
following the existing _-prefix convention for private public-facing
types (e.g. sorted_list/types.ts's _MergeReduce).
