--- fjs/sul/level/hash/module.f.ts
@@ -1,65 +1,0 @@
-/**
- * Streaming encoder for SUL hash-level symbols using a Patricia trie and 256-bit content-addressed hashes.
- *
- * @module
- */
-
-import type { Create, State } from '../../../types/patricia_trie/types.ts'
-import { emptyState, patriciaTrie } from '../../../types/patricia_trie/module.f.mjs'
-import { compress } from '../../id/module.f.mjs'
-import type { Id } from '../../id/types.ts'
-import { asBase } from '../../../types/nominal/module.f.mjs'
-
-/**
- * Called once per merge during encoding. `merged = compress(left, right)`.
- * `isSymbol` is `true` for the terminal `compress(root, t)` that produces the
- * word-level output symbol, and `false` for Patricia trie internal merges.
- * Implementations record the triple in a content-addressed store.
- */
-export type Add<S> = (left: Id, right: Id, merged: Id, isSymbol: boolean, storage: S) => S
-
-/**
- * Streaming state for hash-level encoding.
- * Wraps the Patricia trie state that accumulates the strictly-decreasing
- * prefix of the word currently being encoded.
- */
-export type EncodeState<S> = State<S, Id>
-
-/**
- * Returns a streaming encoder for hash-level symbols.
- *
- * The returned step function processes one level-3 symbol at a time.
- * It returns `undefined` while the strictly-decreasing prefix is still
- * being accumulated. When the terminating symbol `t` arrives (`t >= last`),
- * the word `[s0 > ... > sk, t]` is finalised:
- *
- * 1. The Patricia trie of `[s0, ..., sk]` is collapsed to a root hash.
- * 2. The root is merged with `t` via `compress`, producing the output symbol.
- * 3. State is reset to an empty stack (storage is preserved).
- *
- * `add` is called once for every `compress` call. `isSymbol` is `false` for
- * Patricia trie internal merges and `true` for the terminal `compress(root, t)`.
- */
-export const encode =
-    <S>(add: Add<S>): (symbol: Id, state: EncodeState<S>) => readonly[Id|undefined, EncodeState<S>] =>
-{
-    const create = (isSymbol: boolean): Create<S, Id> => (a, b, s) => {
-        const m = compress(a, b)
-        return [m, add(a, b, m, isSymbol, s)]
-    }
-    const { push, end } = patriciaTrie(create(false))
-    const rootCreate = create(true)
-    return (symbol, state) => {
-        const [, stack] = state
-        const last = stack.at(-1)
-        if (last === undefined || last[0] > asBase(symbol)) {
-            return [undefined, push([asBase(symbol), symbol], state)]
-        }
-        const [root1, storage1] = end(state)
-        const [root2, storage2] = rootCreate(root1!, symbol, storage1)
-        return [root2, [storage2, []]]
-    }
-}
-
-/** Returns the initial encoding state wrapping `storage`. */
-export const emptyEncodeState: <S>(storage: S) => EncodeState<S> = emptyState
