+++ fjs/sul/module.f.ts
@@ -7,7 +7,8 @@
 
 import { emptyPipelineState, pipelineStep } from './level/literal/module.f.mjs'
 import type { PipelineState } from './level/literal/types.ts'
-import { encode as hashEncode, type Add } from './level/hash/module.f.ts'
+import { encode as hashEncode } from './level/hash/module.f.mjs'
+import type { Add } from './level/hash/types.ts'
 import { level3Id } from './id/module.f.mjs'
 import type { Id } from './id/types.ts'
 import type { InternalState } from '../types/patricia_trie/types.ts'
